simplify ClosureGradScaler
