+++ composer/trainer/scaler.py
@@ -7,6 +7,8 @@ import torch
 from torch.cuda.amp.grad_scaler import GradScaler, OptState, _refresh_per_optimizer_state
 from torch.optim import Optimizer
 
+from composer.utils import dist
+
 from composer.core.types import Tensor
 
 
@@ -77,9 +79,9 @@ class ClosureGradScaler(GradScaler):
             retval: float = closure(**kwargs)
 
             should_continue = self._unscale_grads_and_continue(optimizer)
-            should_continue = self.ddp_reduce_scalar_and(should_continue)
+            other_should_continue = dist.all_gather_object(should_continue)
 
-            return retval if should_continue else None
+            return retval if all(other_should_continue) else None
 
         return optimizer.step(closure=_amp_closure)  # type: ignore
 
@@ -136,7 +138,7 @@ class ClosureGradScaler(GradScaler):
                     found_inf_combined += found_infs[i]
 
             # This is the only line changed from original grad_scaler implementation
-            found_inf_combined = self.ddp_reduce_tensor_sum(found_inf_combined)
+            dist.all_reduce(found_inf_combined, reduce_operation="SUM")
 
             torch._amp_update_scale_(_scale, _growth_tracker, found_inf_combined, self._growth_factor,
                                      self._backoff_factor, self._growth_interval)
